requirements.txt
aiohttp[speedups]
aiohttp-swagger[performance]
cx-Oracle
PyYAML
cryptography
loguru
